Fixes for minmax values in GUI and for AI
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@3348 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/server/CHideAndSeek.cpp
@@ -242,16 +242,16 @@ bool CHideAndSeek::CanSee(CWorm* worm1, CWorm* worm2)
 	if(worm1->getTeam() == SEEKER)
 	{
 		if(type & PX_EMPTY)
-			return dist.GetLength() < (float)tLXOptions->tGameInfo.features[FT_HS_SeekerVisionRange];
+			return dist.GetLength() < (int)tLXOptions->tGameInfo.features[FT_HS_SeekerVisionRange];
 		else
-			return dist.GetLength() < (float)tLXOptions->tGameInfo.features[FT_HS_SeekerVisionRangeThroughWalls];
+			return dist.GetLength() < (int)tLXOptions->tGameInfo.features[FT_HS_SeekerVisionRangeThroughWalls];
 	}
 	else 
 	{
 		if(type & PX_EMPTY)
-			return dist.GetLength() < (float)tLXOptions->tGameInfo.features[FT_HS_HiderVisionRange];
+			return dist.GetLength() < (int)tLXOptions->tGameInfo.features[FT_HS_HiderVisionRange];
 		else
-			return dist.GetLength() < (float)tLXOptions->tGameInfo.features[FT_HS_HiderVisionRangeThroughWalls];
+			return dist.GetLength() < (int)tLXOptions->tGameInfo.features[FT_HS_HiderVisionRangeThroughWalls];
 	}
 }
